# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIRAMA API",
        "description": "API untuk aplikasi mobile SIRAMA (Sistem Informasi Administrasi Kependudukan). Autentikasi menggunakan Laravel Sanctum Bearer Token.",
        "contact": {
            "email": "admin@sirama.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sirama.tunggulmajid.my.id/api",
            "description": "SIRAMA API Server (Production)"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "SIRAMA API Server (Local)"
        }
    ],
    "paths": {
...
        "/pengajuan/kk-penambahan": {
            "post": {
                "tags": [
                    "Pengajuan - KK Penambahan"
                ],
                "summary": "Buat pengajuan penambahan anggota KK",
                "description": "Membuat pengajuan penambahan anggota KK sekaligus mengisi form dan mengupload semua dokumen dalam satu request.",
                "operationId": "316dbd13f8213e7df7dd4ea421b36c30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_kepala_keluarga",
                                    "nomor_kk",
                                    "nama_ketua_rt",
                                    "nama_ketua_rw",
                                    "nama_lengkap_tambahan",
                                    "jenis_kelamin_tambahan",
                                    "tempat_lahir_tambahan",
                                    "tanggal_lahir_tambahan",
                                    "status_hubungan",
                                    "kelainan_fisik_mental",
                                    "penyandang_cacat",
                                    "agama",
                                    "nama_ibu_kandung",
                                    "nik_ibu",
                                    "nama_ayah_kandung",
                                    "nik_ayah",
                                    "file_kk_asli",
                                    "file_sk_lahir_akta",
                                    "file_ktp_suami_istri",
                                    "file_surat_nikah"
                                ],
                                "properties": {
                                    "nama_kepala_keluarga": {
                                        "type": "string"
                                    },
                                    "nomor_kk": {
                                        "type": "string"
                                    },
                                    "nama_ketua_rt": {
                                        "type": "string"
                                    },
                                    "nama_ketua_rw": {
                                        "type": "string"
                                    },
                                    "nama_lengkap_tambahan": {
                                        "type": "string"
                                    },
                                    "jenis_kelamin_tambahan": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ]
                                    },
                                    "tempat_lahir_tambahan": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir_tambahan": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status_hubungan": {
                                        "type": "string"
                                    },
                                    "kelainan_fisik_mental": {
                                        "type": "string"
                                    },
                                    "penyandang_cacat": {
                                        "type": "string"
                                    },
                                    "agama": {
                                        "type": "string"
                                    },
                                    "nama_ibu_kandung": {
                                        "type": "string"
                                    },
                                    "nik_ibu": {
                                        "type": "string"
                                    },
                                    "nama_ayah_kandung": {
                                        "type": "string"
                                    },
                                    "nik_ayah": {
                                        "type": "string"
                                    },
                                    "file_kk_asli": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_sk_lahir_akta": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_suami_istri": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_surat_nikah": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengajuan KK Penambahan berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pengajuan/kk-penambahan/{id}": {
            "post": {
                "tags": [
                    "Pengajuan - KK Penambahan"
                ],
                "summary": "Revisi pengajuan KK Penambahan",
                "description": "Mengupdate data form dan/atau dokumen pengajuan KK Penambahan. File yang tidak dikirim tidak akan diubah.",
                "operationId": "d53dedf87c6ffeb810bf0d38345f7bc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama_kepala_keluarga": {
                                        "type": "string"
                                    },
                                    "nomor_kk": {
                                        "type": "string"
                                    },
                                    "nama_ketua_rt": {
                                        "type": "string"
                                    },
                                    "nama_ketua_rw": {
                                        "type": "string"
                                    },
                                    "nama_lengkap_tambahan": {
                                        "type": "string"
                                    },
                                    "jenis_kelamin_tambahan": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ]
                                    },
                                    "tempat_lahir_tambahan": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir_tambahan": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status_hubungan": {
                                        "type": "string"
                                    },
                                    "kelainan_fisik_mental": {
                                        "type": "string"
                                    },
                                    "penyandang_cacat": {
                                        "type": "string"
                                    },
                                    "agama": {
                                        "type": "string"
                                    },
                                    "nama_ibu_kandung": {
                                        "type": "string"
                                    },
                                    "nik_ibu": {
                                        "type": "string"
                                    },
                                    "nama_ayah_kandung": {
                                        "type": "string"
                                    },
                                    "nik_ayah": {
                                        "type": "string"
                                    },
                                    "file_kk_asli": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_sk_lahir_akta": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_suami_istri": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_surat_nikah": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengajuan KK Penambahan berhasil diupdate"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum. Contoh: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}